bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
PyYAML>=6.0.1
pydantic>=2.4.2
PyNaCl>=1.5.0
lavalink>=5.9.0
redis>=5.0.1
prometheus-client>=0.20.0
pyinstrument>=5.1.1
aiohttp>=3.11.2
